+++ filter/bloom/bloom_generic.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Blockwatch Data Inc.
 // Author: stefan@blockwatch.cc
 
-package bloomVec
+package bloom
 
 import (
 	"blockwatch.cc/knoxdb/hash/xxhashVec"
