+++ filter/bloom/bloom_test.go
@@ -1,156 +1,334 @@
-package bloom_test
+// Copyright (c) 2021 Blockwatch Data Inc.
+// Author: stefan@blockwatch.cc
+
+package bloom
 
 import (
 	"encoding/binary"
 	"fmt"
-	"os"
 	"testing"
-
-	"blockwatch.cc/knoxdb/filter/bloom"
 )
 
 // Ensure filter can insert values and verify they exist.
-func TestFilter_InsertContains(t *testing.T) {
-	if testing.Short() || os.Getenv("GORACE") != "" || os.Getenv("APPVEYOR") != "" {
-		t.Skip("Skipping test in short, race and appveyor mode.")
-	}
-
-	// Short, less comprehensive test.
-	testShortFilter_InsertContains(t)
-
+func TestFilterAddContains(t *testing.T) {
+	var num, fsize int
 	if testing.Short() {
-		return // Just run the above short test
+		num = 100000
+		fsize = 1437758
+	} else {
+		num = 10000000
+		fsize = 143775876
 	}
-
-	// More comprehensive test for the xxhash based Bloom Filter.
 
 	// These parameters will result, for 10M entries, with a bloom filter
 	// with 0.001 false positive rate (1 in 1000 values will be incorrectly
 	// identified as being present in the set).
-	filter := bloom.NewFilter(143775876, 10)
+	filter := NewFilter(fsize)
 	v := make([]byte, 4)
-	for i := 0; i < 10000000; i++ {
+	for i := 0; i < num; i++ {
 		binary.BigEndian.PutUint32(v, uint32(i))
 		filter.Add(v)
 	}
 
 	// None of the values inserted should ever be considered "not possibly in
 	// the filter".
-	t.Run("100M", func(t *testing.T) {
-		for i := 0; i < 10000000; i++ {
-			binary.BigEndian.PutUint32(v, uint32(i))
-			if !filter.Contains(v) {
-				t.Fatalf("got false for value %q, expected true", v)
-			}
+	for i := 0; i < num; i++ {
+		binary.BigEndian.PutUint32(v, uint32(i))
+		if !filter.Contains(v) {
+			t.Fatalf("got false for value %q, expected true", v)
 		}
+	}
 
-		// If we check for 100,000,000 values that we know are not present in the
-		// filter then we might expect around 100,000 of them to be false positives.
-		var fp int
-		for i := 10000000; i < 110000000; i++ {
-			binary.BigEndian.PutUint32(v, uint32(i))
-			if filter.Contains(v) {
-				fp++
-			}
+	// If we check for 100,000,000 values that we know are not present in the
+	// filter then we might expect around 100,000 of them to be false positives.
+	var fp int
+	for i := num; i < 11*num; i++ {
+		binary.BigEndian.PutUint32(v, uint32(i))
+		if filter.Contains(v) {
+			fp++
 		}
+	}
 
-		if fp > 1000000 {
-			// If we're an order of magnitude off, then it's arguable that there
-			// is a bug in the bloom filter.
-			t.Fatalf("got %d false positives which is an error rate of %f, expected error rate <=0.001", fp, float64(fp)/100000000)
-		}
-		t.Logf("Bloom false positive error rate was %f", float64(fp)/100000000)
-	})
+	if fp > num/10 {
+		// If we're an order of magnitude off, then it's arguable that there
+		// is a bug in the bloom filter.
+		t.Fatalf("got %d false positives which is an error rate of %f, expected error rate <=0.001", fp, float64(fp)/100000000)
+	}
+	t.Logf("Bloom false positive error rate was %f", float64(fp)/float64(num)/10)
+
 }
 
-func testShortFilter_InsertContains(t *testing.T) {
-	t.Run("short", func(t *testing.T) {
-		f := bloom.NewFilter(1000, 4)
+// Ensure filter can insert values and verify they exist.
+func TestFilterAddContainsUint32Generic(t *testing.T) {
+	var num, fsize int
+	if testing.Short() {
+		num = 100000
+		fsize = 1437758
+	} else {
+		num = 10000000
+		fsize = 143775876
+	}
 
-		// Insert value and validate.
-		f.Add([]byte("Bess"))
-		if !f.Contains([]byte("Bess")) {
-			t.Fatal("expected true")
-		}
-		h1 := bloom.Hash([]byte("Bess"))
-		if !f.ContainsHash(h1) {
-			t.Fatal("expected true")
-		}
+	// These parameters will result, for 10M entries, with a bloom filter
+	// with 0.001 false positive rate (1 in 1000 values will be incorrectly
+	// identified as being present in the set).
+	filter := NewFilter(fsize)
+	slice := make([]uint32, num)
+	for i := 0; i < num; i++ {
+		slice[i] = uint32(i)
+	}
+	filterAddManyUint32Generic(*filter, slice, xxHash32Seed)
 
-		// Insert another value and test.
-		f.Add([]byte("Emma"))
-		if !f.Contains([]byte("Emma")) {
-			t.Fatal("expected true")
+	// None of the values inserted should ever be considered "not possibly in
+	// the filter".
+	for i := 0; i < num; i++ {
+		if !filter.ContainsUint32(uint32(i)) {
+			t.Fatalf("got false for value %v, expected true", i)
 		}
-		h2 := bloom.Hash([]byte("Emma"))
-		if !f.ContainsHash(h2) {
-			t.Fatal("expected true")
-		}
+	}
 
-		// Validate that a non-existent value doesn't exist.
-		if f.Contains([]byte("Jane")) {
-			t.Fatal("expected false")
+	// If we check for 100,000,000 values that we know are not present in the
+	// filter then we might expect around 100,000 of them to be false positives.
+	var fp int
+	for i := num; i < 11*num; i++ {
+		if filter.ContainsUint32(uint32(i)) {
+			fp++
 		}
-		h3 := bloom.Hash([]byte("Jane"))
-		if f.ContainsHash(h3) {
-			t.Fatal("expected false")
-		}
+	}
 
-		// Validate multi-hash
-		if !f.ContainsAnyHash([][2]uint64{h1, h3}) {
-			t.Fatal("expected true")
-		}
-		if !f.ContainsAnyHash([][2]uint64{h2, h3}) {
-			t.Fatal("expected true")
-		}
-		if f.ContainsAnyHash([][2]uint64{h3, h3}) {
-			t.Fatal("expected false")
-		}
+	if fp > num/10 {
+		// If we're an order of magnitude off, then it's arguable that there
+		// is a bug in the bloom filter.
+		t.Fatalf("got %d false positives which is an error rate of %f, expected error rate <=0.001", fp, float64(fp)/100000000)
+	}
+	t.Logf("Bloom false positive error rate was %f", float64(fp)/float64(num)/10)
+}
 
-	})
+// Ensure filter can insert values and verify they exist.
+func TestFilterAddContainsInt32Generic(t *testing.T) {
+	var num, fsize int
+	if testing.Short() {
+		num = 100000
+		fsize = 1437758
+	} else {
+		num = 10000000
+		fsize = 143775876
+	}
+
+	// These parameters will result, for 10M entries, with a bloom filter
+	// with 0.001 false positive rate (1 in 1000 values will be incorrectly
+	// identified as being present in the set).
+	filter := NewFilter(fsize)
+	slice := make([]int32, num)
+	for i := 0; i < num; i++ {
+		slice[i] = int32(i)
+	}
+	filterAddManyInt32Generic(*filter, slice, xxHash32Seed)
+
+	// None of the values inserted should ever be considered "not possibly in
+	// the filter".
+	for i := 0; i < num; i++ {
+		if !filter.ContainsInt32(int32(i)) {
+			t.Fatalf("got false for value %v, expected true", i)
+		}
+	}
+
+	// If we check for 100,000,000 values that we know are not present in the
+	// filter then we might expect around 100,000 of them to be false positives.
+	var fp int
+	for i := num; i < 11*num; i++ {
+		if filter.ContainsInt32(int32(i)) {
+			fp++
+		}
+	}
+
+	if fp > num/10 {
+		// If we're an order of magnitude off, then it's arguable that there
+		// is a bug in the bloom filter.
+		t.Fatalf("got %d false positives which is an error rate of %f, expected error rate <=0.001", fp, float64(fp)/100000000)
+	}
+	t.Logf("Bloom false positive error rate was %f", float64(fp)/float64(num)/10)
+}
+
+// Ensure filter can insert values and verify they exist.
+func TestFilterAddContainsUint64Generic(t *testing.T) {
+	var num, fsize int
+	if testing.Short() {
+		num = 100000
+		fsize = 1437758
+	} else {
+		num = 10000000
+		fsize = 143775876
+	}
+
+	// These parameters will result, for 10M entries, with a bloom filter
+	// with 0.001 false positive rate (1 in 1000 values will be incorrectly
+	// identified as being present in the set).
+	filter := NewFilter(fsize)
+	slice := make([]uint64, num)
+	for i := 0; i < num; i++ {
+		slice[i] = uint64(i)
+	}
+	filterAddManyUint64Generic(*filter, slice, xxHash32Seed)
+
+	// None of the values inserted should ever be considered "not possibly in
+	// the filter".
+	for i := 0; i < num; i++ {
+		if !filter.ContainsUint64(uint64(i)) {
+			t.Fatalf("got false for value %v, expected true", i)
+		}
+	}
+
+	// If we check for 100,000,000 values that we know are not present in the
+	// filter then we might expect around 100,000 of them to be false positives.
+	var fp int
+	for i := num; i < 11*num; i++ {
+		if filter.ContainsUint64(uint64(i)) {
+			fp++
+		}
+	}
+
+	if fp > num/10 {
+		// If we're an order of magnitude off, then it's arguable that there
+		// is a bug in the bloom filter.
+		t.Fatalf("got %d false positives which is an error rate of %f, expected error rate <=0.001", fp, float64(fp)/100000000)
+	}
+	t.Logf("Bloom false positive error rate was %f", float64(fp)/float64(num)/10)
+}
+
+// Ensure filter can insert values and verify they exist.
+func TestFilterAddContainsInt64Generic(t *testing.T) {
+	var num, fsize int
+	if testing.Short() {
+		num = 100000
+		fsize = 1437758
+	} else {
+		num = 10000000
+		fsize = 143775876
+	}
+
+	// These parameters will result, for 10M entries, with a bloom filter
+	// with 0.001 false positive rate (1 in 1000 values will be incorrectly
+	// identified as being present in the set).
+	filter := NewFilter(fsize)
+	slice := make([]int64, num)
+	for i := 0; i < num; i++ {
+		slice[i] = int64(i)
+	}
+	filterAddManyInt64Generic(*filter, slice, xxHash32Seed)
+
+	// None of the values inserted should ever be considered "not possibly in
+	// the filter".
+	for i := 0; i < num; i++ {
+		if !filter.ContainsInt64(int64(i)) {
+			t.Fatalf("got false for value %v, expected true", i)
+		}
+	}
+
+	// If we check for 100,000,000 values that we know are not present in the
+	// filter then we might expect around 100,000 of them to be false positives.
+	var fp int
+	for i := num; i < 11*num; i++ {
+		if filter.ContainsInt64(int64(i)) {
+			fp++
+		}
+	}
+
+	if fp > num/10 {
+		// If we're an order of magnitude off, then it's arguable that there
+		// is a bug in the bloom filter.
+		t.Fatalf("got %d false positives which is an error rate of %f, expected error rate <=0.001", fp, float64(fp)/100000000)
+	}
+	t.Logf("Bloom false positive error rate was %f", float64(fp)/float64(num)/10)
+}
+
+func TestFilterMergeGeneric(t *testing.T) {
+	var num, fsize int
+	if testing.Short() {
+		num = 100000
+		fsize = 1437758
+	} else {
+		num = 10000000
+		fsize = 143775876
+	}
+
+	// These parameters will result, for 10M entries, with a bloom filter
+	// with 0.001 false positive rate (1 in 1000 values will be incorrectly
+	// identified as being present in the set).
+	filter := NewFilter(fsize)
+	slice := make([]uint32, num/2)
+	for i := 0; i < num/2; i++ {
+		slice[i] = uint32(i)
+	}
+	filterAddManyUint32Generic(*filter, slice, xxHash32Seed)
+
+	filter2 := NewFilter(fsize)
+	for i := num / 2; i < num; i++ {
+		slice[i-num/2] = uint32(i)
+	}
+	filterAddManyUint32Generic(*filter2, slice, xxHash32Seed)
+
+	filterMergeGeneric(filter.b, filter2.b)
+
+	// None of the values inserted should ever be considered "not possibly in
+	// the filter".
+	for i := 0; i < num; i++ {
+		if !filter.ContainsUint32(uint32(i)) {
+			t.Fatalf("got false for value %v, expected true", i)
+		}
+	}
+
+	// If we check for 100,000,000 values that we know are not present in the
+	// filter then we might expect around 100,000 of them to be false positives.
+	var fp int
+	for i := num; i < 11*num; i++ {
+		if filter.ContainsUint32(uint32(i)) {
+			fp++
+		}
+	}
+
+	if fp > num/10 {
+		// If we're an order of magnitude off, then it's arguable that there
+		// is a bug in the bloom filter.
+		t.Fatalf("got %d false positives which is an error rate of %f, expected error rate <=0.001", fp, float64(fp)/100000000)
+	}
+	t.Logf("Bloom false positive error rate was %f", float64(fp)/float64(num)/10)
 }
 
 var benchCases = []struct {
-	m, k uint64
-	n    int
-	v    int
+	m int
+	n int
+	v int
 }{
-	// {m: 100, k: 4, n: 1000, v: 1000},
-	// {m: 1000, k: 4, n: 1000, v: 1000},
-	// {m: 10000, k: 4, n: 1000, v: 1000},
-	// {m: 100000, k: 4, n: 1000, v: 1000},
-	// {m: 100, k: 8, n: 1000, v: 1000},
-	// {m: 1000, k: 8, n: 1000, v: 1000},
-	// {m: 10000, k: 8, n: 1000, v: 1000},
-	// {m: 100000, k: 8, n: 1000, v: 1000},
-	// {m: 100, k: 20, n: 1000, v: 1000},
-	// {m: 1000, k: 20, n: 1000, v: 1000},
-	// {m: 10000, k: 20, n: 1000, v: 1000},
-	// {m: 100000, k: 20, n: 1000, v: 1000},
-
 	// 32k packs
-	{m: 524288, k: 11, n: 32768, v: 1}, // 64kB mem, p=0.0005 (optimal)
-	{m: 524288, k: 4, n: 32768, v: 1},  // 64kB mem, p=0.0024
-	{m: 262144, k: 6, n: 32768, v: 1},  // 32kB mem, p=0.022 (optimal)
-	{m: 262144, k: 4, n: 32768, v: 1},  // 32kB mem, p=0.024
+	{m: 32768, n: 32768, v: 1},
+	{m: 65536, n: 32768, v: 1},
+	{m: 131072, n: 32768, v: 1},
+	{m: 262144, n: 32768, v: 1},
+	{m: 524288, n: 32768, v: 1},
+	{m: 1048576, n: 32768, v: 1},
 
 	// 64k packs
-	{m: 1048576, k: 11, n: 65536, v: 1}, // 128k mem, p=0.0005 (optimal)
-	{m: 1048576, k: 4, n: 65536, v: 1},  // 128k mem, p=0.0024
-	{m: 524288, k: 6, n: 65536, v: 1},   // 64k mem,p=0.022 (optimal)
-	{m: 524288, k: 4, n: 65536, v: 1},   // 64k mem,p=0.024
+	{m: 32768, n: 65536, v: 1},
+	{m: 65536, n: 65536, v: 1},
+	{m: 131072, n: 65536, v: 1},
+	{m: 262144, n: 65536, v: 1},
+	{m: 524288, n: 65536, v: 1},
+	{m: 1048576, n: 65536, v: 1},
 }
 
-func BenchmarkFilter_Insert(b *testing.B) {
+func BenchmarkFilterAdd(b *testing.B) {
 	for _, c := range benchCases {
 		data := make([][]byte, 0, c.n)
+		buf := make([]byte, 4)
 		for i := 0; i < c.n; i++ {
-			data = append(data, []byte(fmt.Sprintf("%d", i)))
+			binary.LittleEndian.PutUint32(buf, uint32(i))
+			data = append(data, buf)
 		}
 
-		filter := bloom.NewFilter(c.m, c.k)
-		b.Run(fmt.Sprintf("m=%d_k=%d_n=%d", c.m, c.k, c.n), func(b *testing.B) {
-			b.ReportAllocs()
+		filter := NewFilter(c.m)
+		b.Run(fmt.Sprintf("m=%d_n=%d", c.m, c.n), func(b *testing.B) {
+			b.SetBytes(4 * int64(c.n))
 			for i := 0; i < b.N; i++ {
 				for _, v := range data {
 					filter.Add(v)
@@ -161,7 +339,41 @@ func BenchmarkFilter_Insert(b *testing.B) {
 	}
 }
 
-var okResult bool
+func BenchmarkFilterAddManyUint32Generic(b *testing.B) {
+	for _, c := range benchCases {
+		data := make([]uint32, c.n)
+		for i := 0; i < c.n; i++ {
+			data[i] = uint32(i)
+		}
+
+		filter := NewFilter(c.m)
+		b.Run(fmt.Sprintf("m=%d_n=%d", c.m, c.n), func(b *testing.B) {
+			b.SetBytes(4 * int64(c.n))
+			for i := 0; i < b.N; i++ {
+				filterAddManyUint32Generic(*filter, data, xxHash32Seed)
+			}
+		})
+
+	}
+}
+
+func BenchmarkFilterAddManyUint64Generic(b *testing.B) {
+	for _, c := range benchCases {
+		data := make([]uint64, c.n)
+		for i := 0; i < c.n; i++ {
+			data[i] = uint64(i)
+		}
+
+		filter := NewFilter(c.m)
+		b.Run(fmt.Sprintf("m=%d_n=%d", c.m, c.n), func(b *testing.B) {
+			b.SetBytes(8 * int64(c.n))
+			for i := 0; i < b.N; i++ {
+				filterAddManyUint64Generic(*filter, data, xxHash32Seed)
+			}
+		})
+
+	}
+}
 
 func BenchmarkFilter_Contains(b *testing.B) {
 	for _, c := range benchCases {
@@ -172,12 +384,12 @@ func BenchmarkFilter_Contains(b *testing.B) {
 			notData = append(notData, []byte(fmt.Sprintf("%d", c.n+i)))
 		}
 
-		filter := bloom.NewFilter(c.m, c.k)
+		filter := NewFilter(c.m)
 		for _, v := range data {
 			filter.Add(v)
 		}
 
-		b.Run(fmt.Sprintf("IN m=%d_k=%d_n=%d", c.m, c.k, c.n), func(b *testing.B) {
+		b.Run(fmt.Sprintf("IN m=%d_n=%d", c.m, c.n), func(b *testing.B) {
 			b.ReportAllocs()
 			for i := 0; i < b.N; i++ {
 				for _, v := range data[:c.v] {
@@ -187,7 +399,7 @@ func BenchmarkFilter_Contains(b *testing.B) {
 		})
 
 		// not in
-		b.Run(fmt.Sprintf("NI m=%d_k=%d_n=%d", c.m, c.k, c.n), func(b *testing.B) {
+		b.Run(fmt.Sprintf("NI m=%d_n=%d", c.m, c.n), func(b *testing.B) {
 			b.ReportAllocs()
 			for i := 0; i < b.N; i++ {
 				for _, v := range notData[:c.v] {
@@ -198,30 +410,24 @@ func BenchmarkFilter_Contains(b *testing.B) {
 	}
 }
 
-func BenchmarkFilter_Merge(b *testing.B) {
+func BenchmarkFilterMergeGeneric(b *testing.B) {
 	for _, c := range benchCases {
-		data1 := make([][]byte, 0, c.n)
-		data2 := make([][]byte, 0, c.n)
+		data1 := make([]uint32, c.n)
+		data2 := make([]uint32, c.n)
 		for i := 0; i < c.n; i++ {
-			data1 = append(data1, []byte(fmt.Sprintf("%d", i)))
-			data2 = append(data2, []byte(fmt.Sprintf("%d", c.n+i)))
+			data1[i] = uint32(i)
+			data2[i] = uint32(c.n + i)
 		}
 
-		filter1 := bloom.NewFilter(c.m, c.k)
-		filter2 := bloom.NewFilter(c.m, c.k)
-		for i := 0; i < c.n; i++ {
-			filter1.Add(data1[i])
-			filter2.Add(data2[i])
-		}
+		filter1 := NewFilter(c.m)
+		filter2 := NewFilter(c.m)
+		filter1.AddManyUint32(data1)
+		filter2.AddManyUint32(data2)
 
-		b.Run(fmt.Sprintf("m=%d_k=%d_n=%d", c.m, c.k, c.n), func(b *testing.B) {
-			b.ReportAllocs()
+		b.Run(fmt.Sprintf("m=%d_n=%d", c.m, c.n), func(b *testing.B) {
+			b.SetBytes(int64(c.m >> 3))
 			for i := 0; i < b.N; i++ {
-				other, err := bloom.NewFilterBuffer(filter1.Bytes(), filter1.K())
-				if err != nil {
-					b.Fatal(err)
-				}
-				other.Merge(filter2)
+				filterMergeGeneric(filter1.b, filter2.b)
 			}
 		})
 	}
