+++ filter/filter_test.go
@@ -9,13 +9,14 @@ import (
 	"fmt"
 	"math/rand"
 	"testing"
-	// "time"
 
 	"blockwatch.cc/knoxdb/filter/bloom"
 	"blockwatch.cc/knoxdb/filter/cuckoo"
 	"blockwatch.cc/knoxdb/hash"
 	"blockwatch.cc/knoxdb/vec"
 )
+
+var filterTestSizes = []int{10000}
 
 func randUint64Slice(n, u int) []uint64 {
 	s := make([]uint64, n*u)
@@ -45,7 +46,7 @@ func randBytes(n int) []byte {
 }
 
 func BenchmarkUint64MapFromSorted(B *testing.B) {
-	for _, n := range []int{10, 1000, 10000, 1000000} {
+	for _, n := range filterTestSizes {
 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
 			a := vec.Uint64.Sort(randUint64Slice(n, 1))
 			B.ResetTimer()
@@ -60,58 +61,10 @@ func BenchmarkUint64MapFromSorted(B *testing.B) {
 	}
 }
 
-// func BenchmarkUint64MapFromSortedNoSize(B *testing.B) {
-// 	for _, n := range []int{10, 1000, 10000, 1000000} {
-// 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
-// 			a := util.Uint64Slice(randUint64Slice(n, 1)).Sort()
-// 			B.ResetTimer()
-// 			B.ReportAllocs()
-// 			for i := 0; i < B.N; i++ {
-// 				m := make(map[uint64]struct{})
-// 				for _, v := range a {
-// 					m[v] = struct{}{}
-// 				}
-// 			}
-// 		})
-// 	}
-// }
-
-// func BenchmarkUint64MapFromUnsorted(B *testing.B) {
-// 	for _, n := range []int{10, 1000, 10000, 1000000} {
-// 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
-// 			a := util.Uint64Slice(randUint64Slice(n, 1))
-// 			B.ResetTimer()
-// 			B.ReportAllocs()
-// 			for i := 0; i < B.N; i++ {
-// 				m := make(map[uint64]struct{}, len(a))
-// 				for _, v := range a {
-// 					m[v] = struct{}{}
-// 				}
-// 			}
-// 		})
-// 	}
-// }
-
-// func BenchmarkUint64MapFromUnsortedNoSize(B *testing.B) {
-// 	for _, n := range []int{10, 1000, 10000, 1000000} {
-// 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
-// 			a := util.Uint64Slice(randUint64Slice(n, 1))
-// 			B.ResetTimer()
-// 			B.ReportAllocs()
-// 			for i := 0; i < B.N; i++ {
-// 				m := make(map[uint64]struct{})
-// 				for _, v := range a {
-// 					m[v] = struct{}{}
-// 				}
-// 			}
-// 		})
-// 	}
-// }
-
 // Bytes(32) in hash map
 //
 func BenchmarkBytes32HashMap(B *testing.B) {
-	for _, n := range []int{10, 32, 64, 100, 500, 1000, 10000, 1000000} {
+	for _, n := range filterTestSizes {
 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
 			a := randByteSlice(n, 32)
 			B.ResetTimer()
@@ -131,42 +84,20 @@ func BenchmarkBytes32HashMap(B *testing.B) {
 	}
 }
 
-// func BenchmarkBytes32HashMapNoSize(B *testing.B) {
-// 	for _, n := range []int{10, 1000, 10000, 1000000} {
-// 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
-// 			a := util.ByteSlice(randByteSlice(n, 32))
-// 			B.ResetTimer()
-// 			B.ReportAllocs()
-// 			for i := 0; i < B.N; i++ {
-// 				m := make(map[uint64]struct{})
-// 				for _, v := range a {
-// 					h := hash.NewInlineFNV64a()
-// 					h.Write(v)
-// 					m[h.Sum64()] = struct{}{}
-// 				}
-// 				if got, want := len(m), len(a); got != want {
-// 					B.Errorf("hash collision got=%d want=%d", got, want)
-// 				}
-// 			}
-// 		})
-// 	}
-// }
-
 // Bloom filter on uint64
 //
-const maxFilterError float64 = 0.01
+const maxFilterError float64 = 0.02
 const cuckooFillFactor = 0.75
 const bloomFillFactor = 1
 
 func BenchmarkUint64BloomFromUnsortedLE(B *testing.B) {
-	for _, n := range []int{10, 1000, 10000, 1000000} {
+	for _, n := range filterTestSizes {
 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
 			a := randUint64Slice(n, 1)
 			B.ResetTimer()
 			B.ReportAllocs()
 			for i := 0; i < B.N; i++ {
-				m, k := bloom.Estimate(uint64(float64(len(a))/bloomFillFactor), maxFilterError)
-				filter := bloom.NewFilter(m, k)
+				filter := bloom.NewFilter(n * 4)
 				for _, v := range a {
 					var buf [8]byte
 					binary.LittleEndian.PutUint64(buf[:], v)
@@ -178,14 +109,13 @@ func BenchmarkUint64BloomFromUnsortedLE(B *testing.B) {
 }
 
 func BenchmarkUint64BloomFromSortedLE(B *testing.B) {
-	for _, n := range []int{10, 1000, 10000, 1000000} {
+	for _, n := range filterTestSizes {
 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
 			a := vec.Uint64.Sort(randUint64Slice(n, 1))
 			B.ResetTimer()
 			B.ReportAllocs()
 			for i := 0; i < B.N; i++ {
-				m, k := bloom.Estimate(uint64(float64(len(a))/bloomFillFactor), maxFilterError)
-				filter := bloom.NewFilter(m, k)
+				filter := bloom.NewFilter(n * 4)
 				for _, v := range a {
 					var buf [8]byte
 					binary.LittleEndian.PutUint64(buf[:], v)
@@ -197,14 +127,13 @@ func BenchmarkUint64BloomFromSortedLE(B *testing.B) {
 }
 
 func BenchmarkUint64BloomFromUnsortedBE(B *testing.B) {
-	for _, n := range []int{10, 1000, 10000, 1000000} {
+	for _, n := range filterTestSizes {
 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
 			a := randUint64Slice(n, 1)
 			B.ResetTimer()
 			B.ReportAllocs()
 			for i := 0; i < B.N; i++ {
-				m, k := bloom.Estimate(uint64(float64(len(a))/bloomFillFactor), maxFilterError)
-				filter := bloom.NewFilter(m, k)
+				filter := bloom.NewFilter(n * 4)
 				for _, v := range a {
 					var buf [8]byte
 					binary.BigEndian.PutUint64(buf[:], v)
@@ -216,14 +145,13 @@ func BenchmarkUint64BloomFromUnsortedBE(B *testing.B) {
 }
 
 func BenchmarkUint64BloomFromSortedBE(B *testing.B) {
-	for _, n := range []int{10, 1000, 10000, 1000000} {
+	for _, n := range filterTestSizes {
 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
 			a := vec.Uint64.Sort(randUint64Slice(n, 1))
 			B.ResetTimer()
 			B.ReportAllocs()
 			for i := 0; i < B.N; i++ {
-				m, k := bloom.Estimate(uint64(float64(len(a))/bloomFillFactor), maxFilterError)
-				filter := bloom.NewFilter(m, k)
+				filter := bloom.NewFilter(n * 4)
 				for _, v := range a {
 					var buf [8]byte
 					binary.BigEndian.PutUint64(buf[:], v)
@@ -235,14 +163,13 @@ func BenchmarkUint64BloomFromSortedBE(B *testing.B) {
 }
 
 func BenchmarkBytes32Bloom(B *testing.B) {
-	for _, n := range []int{10, 32, 64, 100, 500, 1000, 10000, 1000000} {
+	for _, n := range filterTestSizes {
 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
 			a := randByteSlice(n, 32)
 			B.ResetTimer()
 			B.ReportAllocs()
 			for i := 0; i < B.N; i++ {
-				m, k := bloom.Estimate(uint64(float64(len(a))/bloomFillFactor), maxFilterError)
-				filter := bloom.NewFilter(m, k)
+				filter := bloom.NewFilter(n * 4)
 				for _, v := range a {
 					filter.Add(v)
 				}
@@ -255,7 +182,7 @@ func BenchmarkBytes32Bloom(B *testing.B) {
 //
 
 func BenchmarkUint64CuckooFromUnsortedLE(B *testing.B) {
-	for _, n := range []int{10, 1000, 10000, 1000000} {
+	for _, n := range filterTestSizes {
 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
 			a := randUint64Slice(n, 1)
 			B.ResetTimer()
@@ -273,7 +200,7 @@ func BenchmarkUint64CuckooFromUnsortedLE(B *testing.B) {
 }
 
 func BenchmarkUint64CuckooFromSortedLE(B *testing.B) {
-	for _, n := range []int{10, 1000, 10000, 1000000} {
+	for _, n := range filterTestSizes {
 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
 			a := vec.Uint64.Sort(randUint64Slice(n, 1))
 			B.ResetTimer()
@@ -291,7 +218,7 @@ func BenchmarkUint64CuckooFromSortedLE(B *testing.B) {
 }
 
 func BenchmarkUint64CuckooFromUnsortedBE(B *testing.B) {
-	for _, n := range []int{10, 1000, 10000, 1000000} {
+	for _, n := range filterTestSizes {
 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
 			a := randUint64Slice(n, 1)
 			B.ResetTimer()
@@ -309,7 +236,7 @@ func BenchmarkUint64CuckooFromUnsortedBE(B *testing.B) {
 }
 
 func BenchmarkUint64CuckooFromSortedBE(B *testing.B) {
-	for _, n := range []int{10, 1000, 10000, 1000000} {
+	for _, n := range filterTestSizes {
 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
 			a := vec.Uint64.Sort(randUint64Slice(n, 1))
 			B.ResetTimer()
@@ -327,7 +254,7 @@ func BenchmarkUint64CuckooFromSortedBE(B *testing.B) {
 }
 
 func BenchmarkBytes32Cuckoo(B *testing.B) {
-	for _, n := range []int{10, 32, 64, 100, 500, 1000, 10000, 1000000} {
+	for _, n := range filterTestSizes {
 		B.Run(fmt.Sprintf("%d", n), func(B *testing.B) {
 			a := randByteSlice(n, 32)
 			B.ResetTimer()
