+++ filter/loglogbeta/LICENSE
@@ -1,3 +1,5 @@
+Original code without vectorization
+
 MIT License
 
 Copyright (c) 2017 Seif Lotfy
