+++ filter/loglogbeta/constants_AVX.h
