+++ filter/loglogbeta/llb.go
@@ -1,7 +1,9 @@
-package llbVec
+// Copyright (c) 2021 Blockwatch Data Inc.
+// Author: stefan@blockwatch.cc
+
+package loglogbeta
 
 import (
-	//	"encoding/binary"
 	"fmt"
 	"math"
 	"math/bits"
