+++ filter/loglogbeta/llbvec_AVX_amd64.s
