+++ filter/loglogbeta/llbvec_amd64.go
@@ -4,7 +4,7 @@
 //go:build go1.7 && amd64 && !gccgo && !appengine
 // +build go1.7,amd64,!gccgo,!appengine
 
-package llbVec
+package loglogbeta
 
 import (
 	"blockwatch.cc/knoxdb/util"
