+++ filter/loglogbeta/llbvec_amd64_test.go
@@ -1,4 +1,7 @@
-package llbVec
+// Copyright (c) 2021 Blockwatch Data Inc.
+// Author: stefan@blockwatch.cc
+
+package loglogbeta
 
 import (
     "fmt"
